refactor: state 간소화 & sealed class 별로 파일 병합
