--- src/main/kotlin/blackjack/domain/state/Blackjack.kt
@@ -1,5 +1,0 @@
-package blackjack.domain.state
-
-import blackjack.domain.Hand
-
-class Blackjack(hand: Hand) : Finished(hand)
