--- src/main/kotlin/blackjack/domain/state/Bust.kt
@@ -1,5 +1,0 @@
-package blackjack.domain.state
-
-import blackjack.domain.Hand
-
-class Bust(hand: Hand) : Finished(hand)
