+++ src/main/kotlin/blackjack/domain/state/Finished.kt
@@ -3,7 +3,7 @@ package blackjack.domain.state
 import blackjack.domain.Card
 import blackjack.domain.Hand
 
-sealed class Finished(hand: Hand) : Started(hand) {
+sealed class Finished(hand: Hand) : Running(hand) {
     override fun draw(card: Card): State {
         throw finishedException()
     }
@@ -16,3 +16,9 @@ sealed class Finished(hand: Hand) : Started(hand) {
 
     private fun finishedException() = IllegalStateException("이미 종료된 상태입니다.")
 }
+
+class Bust(hand: Hand) : Finished(hand)
+
+class Stay(hand: Hand) : Finished(hand)
+
+class Blackjack(hand: Hand) : Finished(hand)
