--- src/main/kotlin/blackjack/domain/state/Started.kt
@@ -1,7 +1,0 @@
-package blackjack.domain.state
-
-import blackjack.domain.Hand
-
-sealed class Started(private val hand: Hand) : State {
-    override fun hand() = hand
-}
