--- src/main/kotlin/blackjack/domain/state/Stay.kt
@@ -1,5 +1,0 @@
-package blackjack.domain.state
-
-import blackjack.domain.Hand
-
-class Stay(hand: Hand) : Finished(hand)
